+++ src/main/java/net/fawnoculus/ntm/blocks/ModBlockEntities.java
@@ -2,6 +2,7 @@ package net.fawnoculus.ntm.blocks;
 
 import net.fabricmc.fabric.api.object.builder.v1.block.entity.FabricBlockEntityTypeBuilder;
 import net.fawnoculus.ntm.blocks.entities.AlloyFurnaceBE;
+import net.fawnoculus.ntm.blocks.entities.RenderTestBE;
 import net.fawnoculus.ntm.main.NTM;
 import net.minecraft.block.Block;
 import net.minecraft.block.entity.BlockEntity;
@@ -12,7 +13,8 @@ import net.minecraft.util.Identifier;
 
 public class ModBlockEntities {
   
-  public static final BlockEntityType<AlloyFurnaceBE> AlloyFurnaceBE = register("alloy_furnace", AlloyFurnaceBE::new, ModBlocks.ALLOY_FURNACE);
+  public static final BlockEntityType<AlloyFurnaceBE> ALLOY_FURNACE_BE = register("alloy_furnace", AlloyFurnaceBE::new, ModBlocks.ALLOY_FURNACE);
+  public static final BlockEntityType<RenderTestBE> RENDER_TEST_BE = register("render_test", RenderTestBE::new, ModBlocks.RENDER_TEST);
   
   private static <T extends BlockEntity> BlockEntityType<T> register(String name, FabricBlockEntityTypeBuilder.Factory<? extends T> entityFactory, Block... blocks) {
     Identifier id = NTM.id(name);