+++ src/main/java/net/fawnoculus/ntm/datagen/ModelProvider.java
@@ -115,7 +115,9 @@ public class ModelProvider extends FabricModelProvider {
     blockStateModelGenerator.registerSimpleCubeAll(ModBlocks.TRIXITE_ORE);
     blockStateModelGenerator.registerSimpleCubeAll(ModBlocks.GEOTHERMAL_VENT);
     blockStateModelGenerator.registerSimpleCubeAll(ModBlocks.BEDROCK_OIL_DEPOSIT);
-    blockStateModelGenerator.registerSimpleCubeAll(ModBlocks.BEDROCK_ORE);
+    
+    
+    blockStateModelGenerator.registerSimpleCubeAll(ModBlocks.RENDER_TEST);
     
     TextureMap alloyFurnaceTextureMap = new TextureMap()
         .put(TextureKey.TOP, TextureMap.getSubId(ModBlocks.ALLOY_FURNACE, "_top"))