+++ Changelog.md
@@ -1,5 +1,6 @@
 # V0.0.2-alpha
 ## Added:
+* Advanced Message System
 * Ore Drops
 * Raw Ores
 * Some Consumables
@@ -11,3 +12,4 @@
 * Alloy Furnace Model
 ## Changed:
 * Mod Description for better Clarity
+* Advanced Tools (they now use the new Message System)